refactor: Use ABCs instead of protocols, move collators into separate module
